[replayer] 0005947: Scipio 200289 - Le replayer plante au démarrage -> merge 4.3.1 20920, 20936

+++ src/libraries/saver_plugin/SaverFacade.cpp
@@ -57,7 +57,7 @@ namespace
 
         virtual void Serialize( tools::OutputBinaryWrapper& output ) const
         {
-            output << buffer_.size();
+            output << static_cast< unsigned int >( buffer_.size() );
             output.Write( buffer_.c_str(), buffer_.size() );
         }
